#  Remove jsonwire.ParseFloat fast path