Calling `strconv.ParseFloat()` does not allocate now, so the necessity of fast path is reduced. This fast path actually slows down the process when it fails.

If overflows, return infinity as all other libs, don't give surprise to user. Since we now always treat overflow as error, this may not be very important, but still a breaking change.

Simple benchmark for your reference
```golang
func BenchmarkParseFloat_NoPoint(b *testing.B) {
	data := []byte("123456789")
	for range b.N {
		ParseFloat(data, 64)
	}
}

func BenchmarkParseFloat_WithPoint(b *testing.B) {
	data := []byte("1234567.89")
	for range b.N {
		ParseFloat(data, 64)
	}
}

func BenchmarkParseFloat_Overflow(b *testing.B) {
	data := []byte("1e1000")
	for range b.N {
		ParseFloat(data, 64)
	}
}
```
before
```
goos: darwin
goarch: arm64
pkg: github.com/go-json-experiment/json/internal/jsonwire
cpu: Apple M2 Pro
BenchmarkParseFloat_NoPoint-12          152670760                7.743 ns/op           0 B/op          0 allocs/op
BenchmarkParseFloat_WithPoint-12        37303958                31.28 ns/op            0 B/op          0 allocs/op
BenchmarkParseFloat_Overflow-12         16789585                69.25 ns/op           56 B/op          2 allocs/op
PASS
ok      github.com/go-json-experiment/json/internal/jsonwire    4.774s
```
after
```
goos: darwin
goarch: arm64
pkg: github.com/go-json-experiment/json/internal/jsonwire
cpu: Apple M2 Pro
BenchmarkParseFloat_NoPoint-12          48795962                24.33 ns/op            0 B/op          0 allocs/op
BenchmarkParseFloat_WithPoint-12        45615379                25.68 ns/op            0 B/op          0 allocs/op
BenchmarkParseFloat_Overflow-12         17682922                65.94 ns/op           56 B/op          2 allocs/op
PASS
ok      github.com/go-json-experiment/json/internal/jsonwire    4.707s
```

Float overflow behavior of other libs:

python3
```
Python 3.13.1 (main, Dec  3 2024, 17:59:52) [Clang 16.0.0 (clang-1600.0.26.4)] on darwin
Type "help", "copyright", "credits" or "license" for more information.
>>> import json
>>> json.loads("1e1000")
inf
```

Chrome
```
> JSON.parse("1e1000")
< Infinity
```