@dsnet 2025-01-21
Thank you for the cleanup!

We can continue to discuss whether `Inf` or `MaxFloat64` is the right value to use in the event of overflow. Should we decide to change behavior, that can be a separate PR.